+++ src/paco/config/paco_context.py
@@ -46,6 +46,17 @@ class AccountContext(object):
         session_cache_filename = '-'.join(['paco', paco_ctx.project.name]) + '.session'
         self.session_cache_path = cache_dir /session_cache_filename
         self.admin_creds = self.paco_ctx.project['credentials']
+
+        # check that account_id has been set
+        # account YAML files are created without an account_id until they are provisioned
+        if self.config.account_id == None:
+            print("""
+The account '{}' is missing an account_id field.
+Add this manually or run `paco provision accounts` for this project.
+""".format(self.config.name)
+            )
+            sys.exit()
+
         self.admin_iam_role_arn = 'arn:aws:iam::{}:role/{}'.format(
             self.config.account_id,
             self.admin_creds.admin_iam_role_name
@@ -237,9 +248,11 @@ class PacoContext(object):
         if account_name in self.accounts:
             return self.accounts[account_name]
 
-        account_ctx = AccountContext(paco_ctx=self,
-                                     name=account_name,
-                                     mfa_account=self.master_account)
+        account_ctx = AccountContext(
+            paco_ctx=self,
+            name=account_name,
+            mfa_account=self.master_account
+        )
         self.accounts[account_name] = account_ctx
 
         return account_ctx
@@ -251,7 +264,7 @@ class PacoContext(object):
             return None
         return region
 
-    def load_project(self, project_init=False, master_only=False):
+    def load_project(self, project_init=False, project_only=False, master_only=False):
         "Load a Paco Project from YAML, initialize settings and controllers, and load Service plug-ins."
         self.project_folder = self.home
         if project_init == True:
@@ -260,6 +273,8 @@ class PacoContext(object):
         # Load the model from YAML
         print("Loading Paco project: %s" % (self.home))
         self.project = load_project_from_yaml(self.project_folder, None)
+        if project_only == True:
+            return
 
         # Settings
         self.build_folder = os.path.join(self.home, "build", self.project.name)
@@ -428,39 +443,6 @@ class PacoContext(object):
                 return valid[answer]
             else:
                 print("Please respond with 'y' or 'n' (or 'yes' or 'no').\n")
-
-    def input(
-        self,
-        prompt,
-        default=None,
-        allowed_values=None,
-        return_bool_on_allowed_value=False,
-        case_sensitive=True
-    ):
-        try_again = True
-        while try_again:
-            suffix = ": "
-            if default != None:
-                suffix += "[%s]: " % str(default)
-            value = input(prompt+suffix) or default
-            if allowed_values != None:
-                for allowed_value in allowed_values:
-                    value_match = False
-                    if isinstance(value, str) and case_sensitive == False:
-                        if allowed_value.lower() == value.lower():
-                            value_match = True
-                    elif allowed_value == value:
-                        value_match = True
-                    if value_match == True:
-                        if return_bool_on_allowed_value == True:
-                            return True
-                        else:
-                            return value
-                print("Invalid value: %s" % (value))
-                print("Allowed values: %s\n" % ', '.join(allowed_values))
-                continue
-            try_again = False
-        return value
 
     def legacy_flag(self, flag):
         if flag in self.project.legacy_flags: